refactoring and getting post in radius

+++ DataPersistence/src/main/java/sep3/project/protobuf/CoordinatesDataOrBuilder.java
@@ -8,14 +8,14 @@ public interface CoordinatesDataOrBuilder extends
     com.google.protobuf.MessageOrBuilder {
 
   /**
-   * <code>float latitude = 1;</code>
+   * <code>double latitude = 1;</code>
    * @return The latitude.
    */
-  float getLatitude();
+  double getLatitude();
 
   /**
-   * <code>float longitude = 2;</code>
+   * <code>double longitude = 2;</code>
    * @return The longitude.
    */
-  float getLongitude();
+  double getLongitude();
 }
